requirements.txt
openai
langchain
langchain-community
langchain-openai
serpapi
google-search-results
gspread
python-dotenv
markdown
Jinja2
linkup-sdk==0.2.5
